requirements.txt
pandas>=2.0
pandas-ta>=0.3.14b0
requests>=2.31
schedule>=1.2.0
python-dotenv>=1.0.0
matplotlib>=3.7
SQLAlchemy>=2.0
alembic>=1.12
psycopg2-binary>=2.9
python-dateutil>=2.8
